+++ spec/event_sourcery/event_processing/esp_process_spec.rb
@@ -14,6 +14,7 @@ RSpec.describe EventSourcery::EventProcessing::ESPProcess do
   let(:subscription_master) { spy(EventSourcery::EventStore::SignalHandlingSubscriptionMaster) }
   let(:error_handler) { double }
   let(:after_fork) { nil }
+  let(:on_event_processor_error) { spy }
 
   describe '#start' do
     subject(:start) { esp_process.start }
@@ -23,6 +24,7 @@ RSpec.describe EventSourcery::EventProcessing::ESPProcess do
       allow(EventSourcery.config.error_handler_class).to receive(:new)
         .with(processor_name: processor_name).and_return(error_handler)
       allow(EventSourcery).to receive(:logger).and_return(logger)
+      allow(EventSourcery.config).to receive(:on_event_processor_error).and_return(on_event_processor_error)
     end
 
     context 'when no error is raised' do
@@ -91,6 +93,7 @@ RSpec.describe EventSourcery::EventProcessing::ESPProcess do
       it 'logs and re-raises the error' do
         expect { start }.to raise_error(error)
         expect(logger).to have_received(:fatal).with(error)
+        expect(on_event_processor_error).to have_received(:call).with(error, processor_name)
       end
     end
   end